Refactor custom QuerySet and Manager classes
Move custom QuerySet and Manager classes.  Add some class methods
referencing other model classes to avoid circular imports.

Use django-model-utils' PassThroughManager for the manager of
the Convictions model to avoid boilerplate proxying methods to
the custom QuerySet.
